Add files via upload
Calculates the motion of volatiles inside the magma ocean in a smarter way using the fact that the coefficient matrix is only updated once every restructuringspacing time steps. Therefore, the eigenvalues&vectors of the matrix and the inverse of the eigenvector (written as a matrix) can be calculated just once every restructuringspacing time steps and reused for all other time steps.