Add files via upload
Smarter calculation of the volatile transport. Since the matrix of coefficients only updates when the planet is restructured, its eigenvalues & vectors can be stored and reused. The coefficients in front of the linear combination of eigenvectors&values that expresses the new masses are different each time step however so they must be calculated. However, they can be calculated by multiplying the previous timestep's volatile masses with the inverse to the matrix of eigenvectors, so we can also have that inverse be precalculated once for every restructuring of the planet. This should together save us time.